Only use the smaller repro case for Runtime_44895.cs
